feat: add video playback control for expanded tile swiper component
