+++ src/libs/components/expanded-tile-swiper/expanded-swiper.loader.ts
@@ -2,7 +2,6 @@ import {
   destroySwiper,
   getActiveSlide,
   getActiveSlideElement,
-  getInstance,
   getSwiperIndexforTile,
   initializeSwiper,
   LookupAttr,
@@ -10,36 +9,20 @@ import {
 } from "../../extensions/swiper/swiper.extension"
 import { waitForElm } from "../../widget.features"
 import { type Swiper } from "swiper"
-import { muteTiktokVideo, pauseTiktokVideo, playTiktokVideo, unMuteTiktokVideo } from "./tiktok-message"
+import { muteTiktokVideo, unMuteTiktokVideo } from "./tiktok-message"
 import { ISdk, SwiperData } from "../../../types"
 import { EVENT_LOAD_MORE } from "../../../events"
 import { getExpandedSlides } from "./base.template"
-import { InvalidMediaError } from "./invalid-media.error"
+import {
+  controlVideoPlayback,
+  controlVideoPlaybackForStory,
+  setupTikTokPlayerReadyEvent,
+  setupVideoEvents,
+  setupYoutubeEvents,
+  YoutubeIframeElementType
+} from "./expanded-tile-video"
 
 declare const sdk: ISdk
-
-type YoutubeContentWindow = Window & {
-  play: () => void
-  pause: () => void
-  reset: () => void
-  mute: () => void
-  unMute: () => void
-}
-
-type YoutubeIframeElementType = HTMLIFrameElement & {
-  contentWindow: YoutubeContentWindow
-}
-
-type SwiperVideoElementType = Window | YoutubeContentWindow | HTMLVideoElement
-
-type VideoSource = "video" | "youtube" | "tiktok"
-
-type SwiperVideoElementData = {
-  element: SwiperVideoElementType
-  source: VideoSource
-}
-
-let tiktokDefaultPlayed = false
 
 /**
  * Initialize/re-initialize swiper for loading expanded tiles
@@ -97,8 +80,8 @@ function initalizeExpandedTile(initialTileId: string, widgetSelector: HTMLElemen
         autoplayTimeLeft: (swiper: Swiper, _timeLeft: number, percentage: number) => {
           storyAutoplayProgress(swiper, percentage)
         },
-        navigationNext: controlVideoPlayback,
-        navigationPrev: controlVideoPlayback
+        navigationNext: swiperNavigationHandler,
+        navigationPrev: swiperNavigationHandler
       }
     },
     getSliderTemplate: getExpandedSlides
@@ -152,10 +135,10 @@ function initalizeStoryExpandedTile(
           storyAutoplayProgress(swiper, percentage)
         },
         navigationNext: async (swiper: Swiper) => {
-          await controlVideoPlaybackForStory(swiper)
+          swiperStoryNavigationHandler(swiper)
         },
         navigationPrev: async (swiper: Swiper) => {
-          await controlVideoPlaybackForStory(swiper)
+          swiperStoryNavigationHandler(swiper)
         },
         autoplay: async (swiper: Swiper) => {
           await controlVideoPlaybackForStory(swiper)
@@ -164,6 +147,32 @@ function initalizeStoryExpandedTile(
     },
     getSliderTemplate: getExpandedSlides
   })
+}
+
+async function swiperStoryNavigationHandler(swiper: Swiper) {
+  const tileId = getTileIdFromSlide(swiper, swiper.realIndex)
+  if (!tileId) {
+    throw Error("Tile ID is not found in next slide from swiper")
+  }
+  const tile = sdk.getTileById(tileId)
+  if (!tile) {
+    throw Error("Tile is not found in next slide from swiper")
+  }
+  sdk.setTile(tile)
+  await controlVideoPlaybackForStory(swiper)
+}
+
+async function swiperNavigationHandler(swiper: Swiper) {
+  const tileId = getTileIdFromSlide(swiper, swiper.realIndex)
+  if (!tileId) {
+    throw Error("Tile ID is not found in next slide from swiper")
+  }
+  const tile = sdk.getTileById(tileId)
+  if (!tile) {
+    throw Error("Tile is not found in next slide from swiper")
+  }
+  sdk.setTile(tile)
+  controlVideoPlayback(swiper)
 }
 
 function storyAutoplayProgress(swiper: Swiper, progress: number) {
@@ -293,198 +302,13 @@ function handleAutoplayProgress(tileWrapper: Element, swiper: Swiper, playCtrl: 
   })
 }
 
-/**
- * Play the video/audio attached to the slide on load where the element is a media element (video/youtube/tiktok)
- */
-async function playMediaOnLoad() {
-  const swiper = getInstance("expanded")
-  if (swiper) {
-    const activeElementData = getSwiperVideoElement(swiper, swiper.realIndex)
-    await triggerPlay(activeElementData)
-  }
-}
+/*
+ Get Tile ID
+*/
 
-function getVideoFromSlide(swiper: Swiper, index: number, isStory = false) {
-  try {
-    return getSwiperVideoElement(swiper, index, isStory)
-  } catch (error) {
-    if (!(error instanceof InvalidMediaError)) throw error
-  }
-  return undefined
-}
-
-/**
- * Play/Pause the video/audio attached to the slide on navigation where the element is a media element (video/youtube/tiktok)
- * @param { Swiper } swiper - the swiper element
- */
-async function controlVideoPlayback(swiper: Swiper) {
-  const activeElement: SwiperVideoElementData | undefined = getVideoFromSlide(swiper, swiper.realIndex)
-  const previousElement: SwiperVideoElementData | undefined = getVideoFromSlide(swiper, swiper.previousIndex)
-
-  if (activeElement) {
-    triggerPlay(activeElement)
-  }
-
-  if (previousElement) {
-    triggerPause(previousElement)
-  }
-}
-
-/**
- * For story widget
- * Play/Pause the video/audio attached to the slide on navigation where the element is a media element (video/youtube/tiktok)
- * @param { Swiper } swiper - the swiper element
- */
-async function controlVideoPlaybackForStory(swiper: Swiper) {
-  const activeElement: SwiperVideoElementData | undefined = getVideoFromSlide(swiper, swiper.realIndex, true)
-  const previousElement: SwiperVideoElementData | undefined = getVideoFromSlide(swiper, swiper.previousIndex, true)
-
-  if (activeElement) {
-    triggerPlay(activeElement)
-  }
-
-  if (previousElement) {
-    triggerPause(previousElement)
-  }
-}
-
-/**
- * Trigger media play for different media element sources ("video", "youtube", "tiktok")
- *
- * @param elementData - the media container element and the source
- * @param elementData.element - the container element of the media (video tag or iframe.contentWindow)
- * @param elementData.source - the media source (video for custom video source, youtube/tiktok)
- */
-async function triggerPlay(elementData?: SwiperVideoElementData) {
-  if (!elementData) {
-    return
-  }
-
-  switch (elementData.source) {
-    case "video": {
-      const videoElement = elementData.element as HTMLVideoElement
-      await videoElement.play()
-      if (window.ugc.swiperContainer["expanded"]?.muted) {
-        videoElement.muted = true
-      } else {
-        videoElement.muted = false
-      }
-      break
-    }
-    case "youtube": {
-      const YoutubeContentWindow = elementData.element as YoutubeContentWindow
-      await YoutubeContentWindow.play()
-      if (window.ugc.swiperContainer["expanded"]?.muted) {
-        YoutubeContentWindow.mute()
-      } else {
-        YoutubeContentWindow.unMute()
-      }
-      break
-    }
-    case "tiktok": {
-      const tiktokFrameWindow = elementData.element as Window
-      await playTiktokVideo(tiktokFrameWindow)
-      if (window.ugc.swiperContainer["expanded"]?.muted) {
-        await muteTiktokVideo(tiktokFrameWindow)
-      } else {
-        await unMuteTiktokVideo(tiktokFrameWindow)
-      }
-      break
-    }
-    default:
-      throw new Error(`unsupported video source ${elementData.source}`)
-  }
-}
-
-/**
- * Trigger media pause for different element sources ("video", "youtube", "tiktok")
- *
- * @param elementData - the media container element and the source
- * @param elementData.element - the container element of the media (video tag or iframe.contentWindow)
- * @param elementData.source - the media source (video for custom video source, youtube/tiktok)
- */
-function triggerPause(elementData?: SwiperVideoElementData) {
-  if (!elementData) {
-    throw new Error("elementData is required")
-  }
-
-  switch (elementData.source) {
-    case "video": {
-      const videoElement = elementData.element as HTMLVideoElement
-      videoElement.pause()
-      videoElement.currentTime = 0
-      break
-    }
-    case "youtube": {
-      const YoutubeContentWindow = elementData.element as YoutubeContentWindow
-      YoutubeContentWindow.pause()
-      YoutubeContentWindow.reset()
-      break
-    }
-    case "tiktok": {
-      const tiktokFrameWindow = elementData.element as Window
-      pauseTiktokVideo(tiktokFrameWindow)
-      break
-    }
-    default:
-      throw new Error(`unsupported video source ${elementData.source}`)
-  }
-}
-
-/**
- * Get swiper video/audio element at the provided index.
- *
- * @param { Swiper } swiper - the swiper element
- * @param { number } index - index of the slide to be returned
- * @param { number } isStory - if it is story widget
- * @returns the video/iframe element or undefined if the element at index is not a video/audio
- */
-function getSwiperVideoElement(swiper: Swiper, index: number, isStory = false): SwiperVideoElementData | undefined {
+export function getTileIdFromSlide(swiper: Swiper, index: number) {
   const element = swiper.slides[index]
-  const tileId = element.getAttribute("data-id")
-  const youtubeId = element.getAttribute("data-yt-id")
-
-  if (!tileId) {
-    throw new Error(`Failed to find tile id for the slide at index ${index}`)
-  }
-
-  const tile = sdk.getTileById(tileId)
-
-  if (!tile) {
-    console.warn(`Failed to find tile data for tile id ${tileId}`)
-  }
-
-  const media = tile?.media
-
-  if (media !== "video" && media !== "short") {
-    return undefined
-  }
-
-  if (youtubeId) {
-    const youtubeFrame = element.querySelector<YoutubeIframeElementType>(`iframe#yt-frame-${tileId}-${youtubeId}`)
-    if (youtubeFrame) {
-      return { element: youtubeFrame.contentWindow, source: "youtube" }
-    }
-  }
-
-  const tiktokId = element.getAttribute("data-tiktok-id")
-
-  if (tiktokId) {
-    const tiktokFrame = element.querySelector<HTMLIFrameElement>(`iframe#tiktok-frame-${tileId}-${tiktokId}`)
-    if (tiktokFrame && tiktokFrame.contentWindow) {
-      return { element: tiktokFrame.contentWindow, source: "tiktok" }
-    }
-  }
-
-  const videoElement = element.querySelector<HTMLVideoElement>(
-    `${isStory ? "" : " .panel .panel-left"} .video-content-wrapper video`
-  )
-
-  if (videoElement) {
-    return { element: videoElement, source: "video" }
-  }
-
-  return undefined
+  return element.getAttribute("data-id")
 }
 
 /**
@@ -623,87 +447,6 @@ function isValidEventSource(sourceId: string) {
 }
 
 /**
- * Setup onload and onerror events for custom video source
- *
- * @param { Element } tile - the tile element
- * @param { HTMLElement } widgetSelector - the container of swiper element
- */
-function setupVideoEvents(tile: Element, widgetSelector: HTMLElement) {
-  const videoSourceElement = tile.querySelector<HTMLVideoElement>("video.video-content > source")
-  if (videoSourceElement) {
-    videoSourceElement.addEventListener("load", () => {
-      playActiveMediaTileOnLoad(tile, widgetSelector)
-    })
-    videoSourceElement.addEventListener("error", () => {
-      videoSourceElement.closest(".video-content-wrapper")?.classList.add("hidden")
-      tile.querySelector(".video-fallback-content")?.classList.remove("hidden")
-    })
-  }
-}
-
-/**
- * Setup onload and onerror (yt-video-error) events for youtube media
- *
- * @param { Element } tile - the tile element
- * @param { HTMLElement } widgetSelector - the container of swiper element
- */
-function setupYoutubeEvents(tile: Element, widgetSelector: HTMLElement) {
-  const tileId = tile.getAttribute("data-id")
-  const youtubeId = tile.getAttribute("data-yt-id")
-
-  if (youtubeId && tileId) {
-    const youtubeFrame = tile.querySelector<HTMLIFrameElement>(`iframe#yt-frame-${tileId}-${youtubeId}`)
-    youtubeFrame?.addEventListener("load", () => {
-      playActiveMediaTileOnLoad(tile, widgetSelector, { name: "data-yt-id", value: youtubeId })
-    })
-    youtubeFrame?.addEventListener("yt-video-error", () => {
-      youtubeFrame.closest(".video-content-wrapper")?.classList.add("hidden")
-      tile.querySelector(".video-fallback-content")?.classList.remove("hidden")
-    })
-  }
-}
-
-/**
- * Setup tiktok player events using window.postMessage api
- * All media are paused by defult and only the media in the active slide is played
- */
-export function setupTikTokPlayerReadyEvent() {
-  tiktokDefaultPlayed = false
-  window.onmessage = (
-    event: MessageEvent<{
-      type: string
-      value?: number
-      "x-tiktok-player": boolean
-    }>
-  ) => {
-    if (event.data["x-tiktok-player"] && event.data.type === "onPlayerReady") {
-      const frameWindow = event.source as Window
-      pauseTiktokVideo(frameWindow)
-
-      if (!tiktokDefaultPlayed) {
-        tiktokDefaultPlayed = true
-        setTimeout(() => playMediaOnLoad(), 300)
-      }
-    }
-  }
-}
-
-/**
- * Play media associated with currently active tile
- *
- * @param { Element } tile - tile element to check
- * @param { HTMLElement } widgetSelector - the container of the swiper element
- * @param { LookupAttr } lookupAttr - additional attribute lookup options for finding the first slide to load
- * @param { LookupAttr.name } lookupAttr.name - name of the attribute for e.g. data-yt-id or data-tiktok-id
- * @param { LookupAttr.value } lookupAttr.value - required value of the attribute
- */
-function playActiveMediaTileOnLoad(tile: Element, widgetSelector: HTMLElement, lookupAttr?: LookupAttr) {
-  if (isActiveTile(tile, widgetSelector, lookupAttr)) {
-    playMediaOnLoad()
-  }
-}
-
-/**
  * Checks if the supplied tile element is currently active in the slide
  *
  * @param { Element } tile - tile element to check
@@ -713,7 +456,7 @@ function playActiveMediaTileOnLoad(tile: Element, widgetSelector: HTMLElement, l
  * @param { LookupAttr.value } lookupAttr.value - required value of the attribute
  * @returns true if the supplied tile element is the currently displayed element. false, otherwise.
  */
-function isActiveTile(tile: Element, widgetSelector: HTMLElement, lookupAttr?: LookupAttr) {
+export function isActiveTile(tile: Element, widgetSelector: HTMLElement, lookupAttr?: LookupAttr) {
   const tileId = tile.getAttribute("data-id")
 
   if (lookupAttr) {
